+++ lib/kettle/gha/pins.rb
@@ -2,13 +2,13 @@
 
 require "version_gem"
 
+require_relative "pins/version_rubric"
 require_relative "pins/version"
 
 module Kettle
   module Gha
     module Pins
       class Error < StandardError; end
-      # Your code goes here...
     end
   end
 end